Added SVG-format rod-stencil
For people using a Cricut machine or printer to print the template on wood, paper, vinyl or other material, this SVG-file can be used. Make sure when aligning the SVG in software and/or printing the file, the square has a size of 50cm (500mm) by 50cm (500mm) for proper spacing between rods.